% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhyloPoissonLogNormal.R
\name{add_replicates}
\alias{add_replicates}
\title{Add replicates to a tree}
\usage{
add_replicates(tree, r)
}
\arguments{
\item{tree}{A phylogenetic tree with n tips.}

\item{r}{the number of replicates too add at each species.}
}
\value{
A phylogenetic tree with n * r tips, and clusters of tips with zero branch lengths.
}
\description{
Utility function to add replicates to a tree, as tips with zero length branches.
}
\keyword{internal}
