% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{check_phyloCompData}
\alias{check_phyloCompData}
\title{Check the validity of a \code{phyloCompData} object}
\usage{
check_phyloCompData(object)
}
\arguments{
\item{object}{A \code{phyloCompData} object}
}
\description{
Check the validity of a \code{phyloCompData} object.
An object that passes the check can be used as the input for the differential expression analysis methods interfaced by \code{compcodeR}.
}
\examples{
mydata <- generateSyntheticData(dataset = "mydata", n.vars = 1000, 
                                samples.per.cond = 5, n.diffexp = 100,
                                id.species = factor(1:10),
                                tree = ape::rphylo(10, 1, 0),
                                lengths.relmeans = "auto", lengths.dispersions = "auto")
check_phyloCompData(mydata)

}
\author{
Charlotte Soneson, Paul Bastide
}
