% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticData.R
\name{computeFactorLengths}
\alias{computeFactorLengths}
\title{Compute Length Normalization Factors}
\usage{
computeFactorLengths(length_matrix, prob.S1, sum.S1)
}
\arguments{
\item{length_matrix}{An n.vars times n.sample matrix of lengths of each gene in each sample.}

\item{prob.S1}{Vector of means for condition 1.}

\item{sum.S1}{Sum of means for condition 1.}
}
\value{
A matrix of the same size as 'length_matrix', with normalization factors
to be applied for each sample and each gene.
}
\description{
Compute the factor to be applied for length normalization.
Each column of the matrix (samples) is normalized by the weighted average of
the column, with weights corresponding to the true probabilities of each gene.
}
\keyword{internal}
