% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{convertListTocompData}
\alias{convertListTocompData}
\title{Convert a list with data and results to a \code{compData} object}
\usage{
convertListTocompData(inp.list)
}
\arguments{
\item{inp.list}{A list with data and results, e.g. generated by \code{compcodeR} version 0.1.0.}
}
\description{
Given a list with data and results (resulting e.g. from \code{compcodeR} version 0.1.0), convert it to a \code{compData} object.
}
\examples{
convertListTocompData(list(count.matrix = matrix(round(1000*runif(4000)), 1000),
                           sample.annotations = data.frame(condition = c(1,1,2,2)),
                           info.parameters = list(dataset = "mydata",
                           uID = "123456")))
}
\author{
Charlotte Soneson
}

