% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhyloPoissonLogNormal.R
\name{deltaStudent}
\alias{deltaStudent}
\title{Multiplying factor for Non central Student}
\usage{
deltaStudent(tree, id.condition, model, selection.strength)
}
\arguments{
\item{tree}{A phylogenetic tree. If \code{NULL}, samples are assumed to be iid.}

\item{id.condition}{A named vector giving the state of each tip (sample).}

\item{model}{The trait evolution model. One of "BM" or "OU".}

\item{selection.strength}{If \code{model="OU"}, the selection strength parameter.}
}
\value{
The multiplying factor.
}
\description{
Under the alternative hypothesis, multiplying factor for the non-central Student
distribution parameter, compared to the case where samples are iid and of equal size.
If the factor is above 1.0, then the problem is "easier" than this the null problem.
If it is below 1.0, then the problem is "harder".
WARNING: This can be identified to the Student distribution only when \code{id.condition} has exactelly two factors.
Otherwise, this is just the square root of the non-central Fisher factor (see \code{\link{deltaFisher}}).
}
\keyword{internal}
