% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylolmMethods.R
\name{extract_results_phylolm}
\alias{extract_results_phylolm}
\title{Extract phylom results}
\usage{
extract_results_phylolm(phylo_lm_obj)
}
\arguments{
\item{phylo_lm_obj}{a phylolm object.}
}
\value{
A list, with:
\describe{
\item{pvalue}{the p value of the differential expression.}
\item{logFC}{the log fold change of the differential expression.}
\item{score}{1 - pvalue.}
}
}
\description{
Extract results from a phylolm object.
The coefficient of interest must be named "condition".
}
\keyword{internal}
