% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticData.R
\name{generateLengths}
\alias{generateLengths}
\title{Simulate a length matrix}
\usage{
generateLengths(id.species, lengths.relmeans, lengths.dispersions)
}
\arguments{
\item{id.species}{An n.sample vector, indicating the species of each sample.}

\item{lengths.relmeans}{A vector of mean values to use in the simulation of
lengths from the Negative Binomial distribution.}

\item{lengths.dispersions}{A vector or matrix of dispersions to use in the
simulation of data from the Negative Binomial distribution.}
}
\value{
A matrix of the same size as 'length_matrix', with normalization
factors to be applied for each sample and each gene.
}
\description{
Simulate a length matrix of size n.vars times n.sample, with the length of
each gene in each sample.
}
\keyword{internal}
