% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylolmMethods.R
\name{getBounds}
\alias{getBounds}
\title{Get Bounds on alpha}
\usage{
getBounds(
  phy,
  factor_up_alpha = 2,
  factor_down_alpha = 3,
  quantile_low_distance = 1e-04
)
}
\arguments{
\item{phy}{a phylogenetic tree.}

\item{factor_up_alpha}{factor for up scalability}

\item{factor_down_alpha}{factor for down scalability}

\item{quantile_low_distance}{quantile for min distance}
}
\value{
a vector with lower and upper bounds for alpha
}
\description{
Return the tree of a \code{phyloCompData} object.
If no tree, return a star tree with unit height, and throw a warning.
}
\details{
If quantile_low_distance=0, then quantile(d_ij)=min(d_ij), and,
for any two tips i,j, the correlation between i and j is bounded by exp(-factor_up_alpha/2).
Those values of alpha will be used for the re-scaling of the tree,
which has an exponential term in exp(2*alpha*h).
The function makes sure that this number is below the maximal float allowed
(equals to .Machine$double.xmax).

See function \code{find_grid_alpha} from package \code{PhylogeneticEM}.
}
\keyword{internal}
