% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eveMethods.R
\name{getColSpecies}
\alias{getColSpecies}
\title{Columns indicators}
\usage{
getColSpecies(cdata, tree_norep)
}
\arguments{
\item{cdata}{a \code{phyloCompData} object}

\item{tree_norep}{a phylogenetic tree without replicates}
}
\value{
vector of names
}
\description{
Get the reference species for each column
}
\keyword{internal}
