% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eveMethods.R
\name{getIsTheta2edge}
\alias{getIsTheta2edge}
\title{Get the edges in theta2 with parsimony}
\usage{
getIsTheta2edge(cdata, tree_norep = getTreeEVE(cdata))
}
\arguments{
\item{cdata}{a \code{phyloCompData} object}

\item{tree_norep}{a phylogenetic tree without replicates}
}
\value{
a vector specifying the regime of each branch of the tree, to be fed
to the \code{isTheta2edge} of function \code{evemodel::fitTwoTheta}
}
\description{
Utility function to paint the edges of the tree from conditions at the tips.
}
\keyword{internal}
