% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eveMethods.R
\name{getlogFCEVE}
\alias{getlogFCEVE}
\title{Log fold change}
\usage{
getlogFCEVE(twoThetaRes, isTheta2edge, tree_norep)
}
\arguments{
\item{twoThetaRes}{result of function \code{evemodel::fitTwoTheta}}

\item{isTheta2edge}{see matching parameter of \code{evemodel::fitTwoTheta}}

\item{tree_norep}{tree used in \code{evemodel::fitTwoTheta}}
}
\value{
vector of logFC
}
\description{
Get the log fold change from a two theta fit with evemodel.
}
\keyword{internal}
