% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticData.R
\name{getNegativeBinomialDispersion}
\alias{getNegativeBinomialDispersion}
\title{Get NB dispersion}
\usage{
getNegativeBinomialDispersion(
  i,
  j,
  S1,
  truedispersions.S1,
  S2,
  truedispersions.S2
)
}
\arguments{
\item{i}{gene indice.}

\item{j}{sample indice.}

\item{S1}{Indices in condition 1.}

\item{truedispersions.S1}{Vector of dispersions for condition 1.}

\item{S2}{Indices in condition 2.}

\item{truedispersions.S2}{Vector of dispersions for condition 2.}
}
\value{
The dispersion for gene i in sample j.
}
\description{
Get the NB dispersion for one gene in one sample
}
\keyword{internal}
