% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticData.R
\name{getNegativeBinomialMean}
\alias{getNegativeBinomialMean}
\title{Get NB mean}
\usage{
getNegativeBinomialMean(
  i,
  j,
  S1,
  prob.S1,
  sum.S1,
  nfact_length.S1,
  S2,
  prob.S2,
  sum.S2,
  nfact_length.S2,
  seq.depths
)
}
\arguments{
\item{i}{gene index.}

\item{j}{sample index.}

\item{S1}{Indices in condition 1.}

\item{prob.S1}{Vector of means for condition 1.}

\item{sum.S1}{Sum of means for condition 1.}

\item{nfact_length.S1}{Matrix of length factors for condition 1.}

\item{S2}{Indices in condition 2.}

\item{prob.S2}{Vector of means for condition 2.}

\item{sum.S2}{Sum of means for condition 2.}

\item{nfact_length.S2}{Matrix of length factors for condition 2.}
}
\value{
The mean for gene i in sample j.
}
\description{
Get the NB mean for one gene in one sample
}
\keyword{internal}
