% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylolmMethods.R
\name{getTree}
\alias{getTree}
\title{Get the tree from a compData object}
\usage{
getTree(cdata)
}
\arguments{
\item{cdata}{a compData object.}
}
\value{
A tree of class \code{phylo}
}
\description{
Return the tree of a \code{phyloCompData} object.
If no tree, return a star tree with unit height, and throw a warning.
}
\keyword{internal}
