% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eveMethods.R
\name{getTreeEVE}
\alias{getTreeEVE}
\title{Get the tree from a phyloCompData object}
\usage{
getTreeEVE(cdata)
}
\arguments{
\item{cdata}{a phyloCompData object.}
}
\value{
A tree of class \code{phylo}
}
\description{
Return the tree of a \code{phyloCompData} object,
removing replicates for use in EVE.
If no tree, return a star tree with unit height, and throw a warning.
}
\keyword{internal}
