% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhyloPoissonLogNormal.R
\name{NB_to_PLN}
\alias{NB_to_PLN}
\title{Negative Binomial to Poisson Log-Normal}
\usage{
NB_to_PLN(mean, dispersion)
}
\arguments{
\item{mean}{mean of the negative binomial.}

\item{dispersion}{dispersion of the negative binomial.}
}
\value{
A list, with:
\describe{
\item{log_means_pln}{Mean of the Poisson log normal in the log space.}
\item{log_variances_pln}{Variance of the Poisson log normal in the log space.}
}
}
\description{
From the parameters of a negative binomial (mean and dispersion),
compute the parameters of a Poisson log-normal with the
same expectation and variance.
}
\keyword{internal}
