% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhyloPoissonLogNormal.R
\name{nEffNaive}
\alias{nEffNaive}
\title{Effective Sample Size}
\usage{
nEffNaive(tree, id.condition, model, selection.strength)
}
\arguments{
\item{tree}{A phylogenetic tree. If \code{NULL}, samples are assumed to be iid.}

\item{id.condition}{A named vector giving the state of each tip (sample).}

\item{model}{The trait evolution model. One of "BM" or "OU".}

\item{selection.strength}{If \code{model="OU"}, the selection strength parameter.}
}
\value{
The effective sample size.
}
\description{
Sample size so that the variance of the estimator is sigma^2 / nEff.
}
\keyword{internal}
