% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{phyloCompDataFromCompData}
\alias{phyloCompDataFromCompData}
\title{Create a \code{phyloCompData} object}
\usage{
phyloCompDataFromCompData(
  compDataObject,
  tree = list(),
  length.matrix = matrix(NA_integer_, 0, 0)
)
}
\arguments{
\item{compDataObject}{An object of class \code{\link{compData}}.}

\item{tree}{A phylogenetic tree describing the relationships between samples.}

\item{length.matrix}{A length matrix, with genes as rows and observations as columns.}
}
\value{
A \code{phyloCompData} object.
}
\description{
The \code{phyloCompData} class extends the \code{\link{compData}} class
with sequence length and phylogeny related information.
}
\author{
Charlotte Soneson, Paul Bastide
}
\keyword{internal}
