% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eveMethods.R
\name{removeReplicatesFromTree}
\alias{removeReplicatesFromTree}
\title{Add replicates to a tree}
\usage{
removeReplicatesFromTree(tree, traits, species = "species", id = "id")
}
\arguments{
\item{tree}{A phylogenetic tree with n tips.}

\item{traits}{A data frame containing at least two columns,
one with sample ids, and on with species names for each samples.}

\item{species}{Name of the column containing species names. Default to "species".}

\item{id}{Name of the column containing samples ids. Default to "id".}
}
\value{
A phylogenetic tree with as many tips as the number of rows in \code{traits},
 and clusters of tips with zero branch lengths corresponding to replicates.
}
\description{
Utility function to add replicates to a tree, as tips with zero length branches.
}
\keyword{internal}
