% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhyloPoissonLogNormal.R
\name{rpois_robust}
\alias{rpois_robust}
\title{Robust rpois}
\usage{
rpois_robust(n, lambda)
}
\arguments{
\item{n}{number of random values to return.}

\item{lambda}{vector of (non-negative) means. Must be of length n.}
}
\value{
A vector of n random draws.
}
\description{
When \code{lambda < 0.999 * .Machine$integer.max}, this function uses the vanilla \code{\link[stats]{rpois}} function.
When lambda is larger, it uses the Gaussian approximation.
}
\details{
Idea comes from: https://stackoverflow.com/questions/23770926/rpois-generates-na-with-large-means-lambda-in-r
}
\keyword{internal}
