% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show,phyloCompData-method}
\alias{show,phyloCompData-method}
\title{Show method for \code{phyloCompData} object}
\usage{
\S4method{show}{phyloCompData}(object)
}
\arguments{
\item{object}{A \code{phyloCompData} object}
}
\description{
Show method for \code{phyloCompData} object.
}
\examples{
mydata <- generateSyntheticData(dataset = "mydata", n.vars = 12500, 
                                samples.per.cond = 5, n.diffexp = 1250)
mydata
}
\author{
Charlotte Soneson, Paul Bastide
}
