% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show_compData}
\alias{show_compData}
\title{Show function for \code{compData} object}
\usage{
show_compData(object)
}
\arguments{
\item{object}{A \code{compData} object}
}
\description{
Show function for \code{compData} object.
}
\keyword{internal}
