% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateRmdCodeDiffExp.R
\name{TCC.createRmd}
\alias{TCC.createRmd}
\title{Generate a \code{.Rmd} file containing code to perform differential expression analysis with TCC}
\usage{
TCC.createRmd(data.path, result.path, codefile, norm.method, test.method,
  iteration = 3, normFDR = 0.1, floorPDEG = 0.05)
}
\arguments{
\item{data.path}{The path to a .rds file containing the \code{compData} object that will be used for the differential expression analysis.}

\item{result.path}{The path to the file where the result object will be saved.}

\item{codefile}{The path to the file where the code will be written.}

\item{norm.method}{The between-sample normalization method used to compensate for varying library sizes and composition in the differential expression analysis. Possible values are \code{"tmm"}, and \code{"deseq"}.}

\item{test.method}{The method used in TCC to find differentially expressed genes. Possible values are \code{"edger"}, \code{"deseq"} and \code{"bayseq"}.}

\item{iteration}{The number of iterations used to find the normalization factors. Default value is 3.}

\item{normFDR}{The FDR cutoff for calling differentially expressed genes in the computation of the normalization factors. Default value is 0.1.}

\item{floorPDEG}{The minimum value to be eliminated as potential differentially expressed genes before performing step 3 in the TCC algorithm. Default value is 0.05.}
}
\description{
A function to generate code that can be run to perform differential expression analysis of RNAseq data (comparing two conditions) using the TCC package. The code is written to a \code{.Rmd} file. This function is generally not called by the user, the main interface for performing differential expression analysis is the \code{\link{runDiffExp}} function.
}
\details{
For more information about the methods and the interpretation of the parameters, see the \code{TCC} package and the corresponding publications.
}
\examples{
try(
if (require(TCC)) {
tmpdir <- normalizePath(tempdir(), winslash = "/")
mydata.obj <- generateSyntheticData(dataset = "mydata", n.vars = 1000,
                                    samples.per.cond = 5, n.diffexp = 100,
                                    output.file = file.path(tmpdir, "mydata.rds"))
runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), result.extent = "TCC",
           Rmdfunction = "TCC.createRmd",
           output.directory = tmpdir, norm.method = "tmm",
           test.method = "edger")
})
}
\author{
Charlotte Soneson
}
\references{
Kadota K, Nishiyama T, and Shimizu K. A normalization strategy for comparing tag count data. Algorithms Mol Biol. 7:5, 2012.

Sun J, Nishiyama T, Shimizu K, and Kadota K. TCC: an R package for comparing tag count data with robust normalization strategies. BMC Bioinformatics 14:219, 2013.
}

