% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhyloPoissonLogNormal.R
\name{vanillaPowerStudent}
\alias{vanillaPowerStudent}
\title{Vanilla Power}
\usage{
vanillaPowerStudent(tree, id.condition, level = 0.95)
}
\arguments{
\item{tree}{A phylogenetic tree. If \code{NULL}, samples are assumed to be iid.}

\item{id.condition}{A named vector giving the state of each tip (sample).}

\item{level}{The level of the t-test.}
}
\value{
The multiplying factor.
}
\description{
Power of the Student t-test between the two conditions for a Gaussian vector
of data, with unit effect size (expectation) and variance.
The tree-induced variance is taken into account.
WARNING: This is only valid when \code{id.condition} has exactelly two factors.
}
\keyword{internal}
