% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhyloPoissonLogNormal.R
\name{vanillaPowerStudentInd}
\alias{vanillaPowerStudentInd}
\title{Vanilla Power for Independent Data}
\usage{
vanillaPowerStudentInd(tree, id.condition, level = 0.95)
}
\arguments{
\item{tree}{A phylogenetic tree. If \code{NULL}, samples are assumed to be iid.}

\item{id.condition}{A named vector giving the state of each tip (sample).}

\item{level}{The level of the t-test.}
}
\value{
The multiplying factor.
}
\description{
Power of the Student t-test between the two conditions for an independent
Gaussian vector of data, with unit effect size (expectation) and variance.
WARNING: This is only valid when \code{id.condition} has exactelly two factors.
}
\keyword{internal}
