% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylolmMethods.R
\name{voom.phylolm.createRmd}
\alias{voom.phylolm.createRmd}
\title{Generate a \code{.Rmd} file containing code to perform differential expression analysis with \code{\link[phylolm]{phylolm}}.}
\usage{
voom.phylolm.createRmd(
  data.path,
  result.path,
  codefile,
  norm.method,
  model = "BM",
  measurement_error = TRUE,
  extraDesignFactors = NULL,
  extraNormFactor = NULL,
  ...
)
}
\arguments{
\item{data.path}{The path to a .rds file containing the \code{compData} object that will be used for the differential expression analysis.}

\item{result.path}{The path to the file where the result object will be saved.}

\item{codefile}{The path to the file where the code will be written.}

\item{norm.method}{The between-sample normalization method used to compensate for varying library sizes and composition in the differential expression analysis. The normalization factors are calculated using the \code{calcNormFactors} of the \code{edgeR} package. Possible values are \code{"TMM"}, \code{"RLE"}, \code{"upperquartile"} and \code{"none"}}

\item{model}{The model for trait evolution on the tree. Default to "BM".}

\item{measurement_error}{A logical value indicating whether there is measurement error. Default to TRUE.}

\item{extraDesignFactors}{A vector containing the extra factors to be passed to the design matrix of \code{DESeq2}. All the factors need to be a \code{sample.annotations} from the \code{\link{compData}} object. It should not contain the "condition" factor column, that will be added automatically.}

\item{extraNormFactor}{A matrix name for normalisation factors, to be applied in addition to the standard ones (see details). The matrix should be one of the \code{info.parameters} list entry of the \code{\link{compData}} object.}

\item{...}{Further arguments to be passed to function \code{\link[phylolm]{phylolm}}.}
}
\value{
The function generates a \code{.Rmd} file containing the code for performing the differential expression analysis. This file can be executed using e.g. the \code{knitr} package.
}
\description{
A function to generate code that can be run to perform differential expression analysis of RNAseq data (comparing two conditions) using the phylolm package. The code is written to a \code{.Rmd} file. This function is generally not called by the user, the main interface for performing differential expression analysis is the \code{\link{runDiffExp}} function.
}
\details{
For more information about the methods and the interpretation of the parameters, see the \code{\link[phylolm]{phylolm}} package and the corresponding publications.



\code{extraNormFactor} is typically the length of the genes (or COGs) for normalization. The count matrix is divided by them before the normalization factor is computed.
}
\examples{
try(
if (require(ape) && require(phylolm)) {
tmpdir <- normalizePath(tempdir(), winslash = "/")
set.seed(20200317)
tree <- rphylo(10, 0.1, 0)
mydata.obj <- generateSyntheticData(dataset = "mydata", n.vars = 1000, 
                                    samples.per.cond = 5, n.diffexp = 100, 
                                    tree = tree,
                                    output.file = file.path(tmpdir, "mydata.rds"))
## Add annotations
sample.annotations(mydata.obj)$test_factor <- factor(rep(1:2, 5))
n_counts <- prod(dim(count.matrix(mydata.obj)))
info.parameters(mydata.obj)$extraNormFactor <- matrix(rpois(n_counts, 100), ncol = 10)
saveRDS(mydata.obj, file.path(tmpdir, "mydata.rds"))
## Diff Exp
runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), result.extent = "DESeq2", 
           Rmdfunction = "DESeq2.norm.createRmd", 
           output.directory = tmpdir,
           norm.method = "TMM",
           extraDesignFactors = c("test_factor", "test_reg"),
           extraNormFactor = c("extraNormFactor"))
           
## Add annotations
sample.annotations(mydata.obj)$test_factor <- factor(rep(1:2, 5))
## Diff Exp
runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), result.extent = "DESeq2", 
           Rmdfunction = "phylolm.createRmd", 
           output.directory = tmpdir, 
           extraDesignFactors = "test_factor", lower.bound = 0.1)
})
}
\references{
Ho, L. S. T. and Ane, C. 2014. "A linear-time algorithm for Gaussian and non-Gaussian trait evolution models". Systematic Biology 63(3):397-408.
}
