% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylolmMethods.R
\name{vst.phylolm.createRmd}
\alias{vst.phylolm.createRmd}
\title{Generate a \code{.Rmd} file containing code to perform differential expression analysis with \code{\link[phylolm]{phylolm}}.}
\usage{
vst.phylolm.createRmd(
  data.path,
  result.path,
  codefile,
  fit.type,
  model = "BM",
  measurement_error = TRUE,
  extraDesignFactors = NULL,
  ...
)
}
\arguments{
\item{data.path}{The path to a .rds file containing the \code{compData} object that will be used for the differential expression analysis.}

\item{result.path}{The path to the file where the result object will be saved.}

\item{codefile}{The path to the file where the code will be written.}

\item{fit.type}{The fitting method used to get the dispersion-mean relationship. Possible values are \code{"parametric"} and \code{"local"}.}

\item{model}{The model for trait evolution on the tree. Default to "BM".}

\item{measurement_error}{A logical value indicating whether there is measurement error. Default to TRUE.}

\item{extraDesignFactors}{A vector containing the extra factors to be passed to the design matrix of \code{DESeq2}. All the factors need to be a \code{sample.annotations} from the \code{\link{compData}} object. It should not contain the "condition" factor column, that will be added automatically.}

\item{...}{Further arguments to be passed to function \code{\link[phylolm]{phylolm}}.}
}
\value{
The function generates a \code{.Rmd} file containing the code for performing the differential expression analysis. This file can be executed using e.g. the \code{knitr} package.
}
\description{
A function to generate code that can be run to perform differential expression analysis of RNAseq data (comparing two conditions) using the phylolm package. The code is written to a \code{.Rmd} file. This function is generally not called by the user, the main interface for performing differential expression analysis is the \code{\link{runDiffExp}} function.
}
\details{
For more information about the methods and the interpretation of the parameters, see the \code{\link[phylolm]{phylolm}} package and the corresponding publications.


The \code{length.matrix} field of the \code{compData}
object is used to normalize the counts. 
\describe{
\item{\code{none}:}{No length normalization.}
\item{\code{TPM}:}{The raw counts are divided by the length of their associated genes before normalization by \code{voom}.}
\item{\code{RPKM}:}{The log2 length is substracted to the log2 CPM computed by \code{voom} for each gene and sample.}
}
}
\examples{
try(
if (require(ape) && require(phylolm)) {
tmpdir <- normalizePath(tempdir(), winslash = "/")
set.seed(20200317)
tree <- rphylo(10, 0.1, 0)
mydata.obj <- generateSyntheticData(dataset = "mydata", n.vars = 1000, 
                                    samples.per.cond = 5, n.diffexp = 100, 
                                    tree = tree,
                                    id.species = 1:10,
                                    lengths.relmeans = rpois(1000, 1000),
                                    lengths.dispersions = rgamma(1000, 1, 1),
                                    output.file = file.path(tmpdir, "mydata.rds"))
## Add annotations
sample.annotations(mydata.obj)$test_factor <- factor(rep(1:2, 5))
saveRDS(mydata.obj, file.path(tmpdir, "mydata.rds"))
## Diff Exp
runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), result.extent = "vst.phylolm", 
           Rmdfunction = "vst.phylolm.createRmd", 
           output.directory = tmpdir,
           fit.type = "parametric",
           extraDesignFactors = c("test_factor", "test_reg"))
})
}
\references{
Ho, L. S. T. and Ane, C. 2014. "A linear-time algorithm for Gaussian and non-Gaussian trait evolution models". Systematic Biology 63(3):397-408.
}
