% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylolmMethods.R
\name{writeNormalization_phyloDE}
\alias{writeNormalization_phyloDE}
\title{Generate a \code{.Rmd} file containing code to normalize data.}
\usage{
writeNormalization_phyloDE(
  norm.method,
  length.normalization,
  data.transformation,
  codefile
)
}
\arguments{
\item{norm.method}{The between-sample normalization method used to compensate for varying library sizes and composition in the differential expression analysis. The normalization factors are calculated using the \code{calcNormFactors} of the \code{edgeR} package. Possible values are \code{"TMM"}, \code{"RLE"}, \code{"upperquartile"} and \code{"none"}}

\item{length.normalization}{one of "none" (no correction), "TPM", "RPKM" (default). See details.}

\item{data.transformation}{one of "log2", "asin(sqrt)" or "sqrt." Data transformation to apply to the normalized data.}

\item{codefile}{}
}
\description{
This function uses \code{\link[phyloDE]{lengthNormalizeRNASeq}}.
}
\keyword{internal}
