\docType{package}
\name{compcodeR-package}
\alias{compcodeR}
\alias{compcodeR-package}
\title{RNAseq data simulation, differential expression analysis and
performance comparison of differential expression methods}
\description{
RNAseq data simulation, differential expression analysis
and performance comparison of differential expression
methods
}
\details{
This package provides extensive functionality for comparing
results obtained by different methods for differential
expression analysis of RNAseq data. It also contains
functions for simulating count data and interfaces to
several packages for performing the differential expression
analysis.
}
\author{
Charlotte Soneson
}
\keyword{package}

