% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{convertphyloCompDataToList}
\alias{convertphyloCompDataToList}
\title{Convert a \code{phyloCompData} object to a list}
\usage{
convertphyloCompDataToList(cpd)
}
\arguments{
\item{cpd}{A \code{phyloCompData} object}
}
\description{
Given a \code{phyloCompData} object, convert it to a list.
}
\examples{
mydata.obj <- generateSyntheticData(dataset = "mydata", n.vars = 1000, 
                                    samples.per.cond = 5, n.diffexp = 100,
                                    id.species = factor(1:10),
                                    tree = ape::rphylo(10, 1, 0))
mydata.list <- convertcompDataToList(mydata.obj)

}
\author{
Charlotte Soneson, Paul Bastide
}
