% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhyloPoissonLogNormal.R
\name{deltaFisher}
\alias{deltaFisher}
\title{Multiplying factor for Non central Fisher}
\usage{
deltaFisher(tree, id.condition)
}
\arguments{
\item{tree}{A phylogenetic tree. If \code{NULL}, samples are assumed to be iid.}

\item{id.condition}{A named vector giving the state of each tip (sample).}
}
\value{
The multiplying factor.
}
\description{
Under the alternative hypothesis, multiplying factor for the non-central Fisher
distribution parameter, compared to the case where samples are iid and of equal size.
If the factor is above 1.0, then the problem is "easier" than this the null problem.
If it is below 1.0, then the problem is "harder".
}
\keyword{internal}
