% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eveMethods.R
\name{evemodel.twoThetaTest.createRmd}
\alias{evemodel.twoThetaTest.createRmd}
\title{Generate a \code{.Rmd} file containing code to perform differential expression analysis with \code{\link[evemodel]{twoThetaTest}}.}
\usage{
evemodel.twoThetaTest.createRmd(
  data.path,
  result.path,
  codefile,
  norm.method,
  length.normalization = "RPKM",
  data.transformation = "log2",
  ...
)
}
\arguments{
\item{data.path}{The path to a .rds file containing the \code{phyloCompData} object that will be used for the differential expression analysis.}

\item{result.path}{The path to the file where the result object will be saved.}

\item{codefile}{The path to the file where the code will be written.}

\item{norm.method}{The between-sample normalization method used to compensate for varying library sizes and composition in the differential expression analysis. The normalization factors are calculated using the \code{calcNormFactors} of the \code{edgeR} package. Possible values are \code{"TMM"}, \code{"RLE"}, \code{"upperquartile"} and \code{"none"}}

\item{length.normalization}{one of "none" (no correction), "TPM" or "RPKM" (default). See details.}

\item{data.transformation}{one of "log2", "asin(sqrt)" or "sqrt". Data transformation to apply to the normalized data.}

\item{...}{Further arguments to be passed to function \code{\link[evemodel]{twoThetaTest}}.}
}
\value{
The function generates a \code{.Rmd} file containing the code for performing the differential expression analysis. This file can be executed using e.g. the \code{knitr} package.
}
\description{
A function to generate code that can be run to perform differential expression analysis of RNAseq data (comparing two conditions) using the evemodel package. The code is written to a \code{.Rmd} file. This function is generally not called by the user, the main interface for performing differential expression analysis is the \code{\link{runDiffExp}} function.
}
\details{
For more information about the methods and the interpretation of the parameters, see the \code{\link[evemodel]{twoThetaTest}} package and the corresponding publications.


The \code{length.matrix} field of the \code{phyloCompData} object 
is used to normalize the counts, using one of the following formulas:
* \code{length.normalization="none"} : \eqn{CPM_{gi} = \frac{N_{gi} + 0.5}{NF_i \times \sum_{g} N_{gi} + 1} \times 10^6}
* \code{length.normalization="TPM"} : \eqn{TPM_{gi} = \frac{(N_{gi} + 0.5) / L_{gi}}{NF_i \times \sum_{g} N_{gi}/L_{gi} + 1} \times 10^6}
* \code{length.normalization="RPKM"} : \eqn{RPKM_{gi} = \frac{(N_{gi} + 0.5) / L_{gi}}{NF_i \times \sum_{g} N_{gi} + 1} \times 10^9}

where \eqn{N_{gi}} is the count for gene g and sample i,
where \eqn{L_{gi}} is the length of gene g in sample i,
and \eqn{NF_i} is the normalization for sample i,
normalized using \code{calcNormFactors} of the \code{edgeR} package.

The function specified by the \code{data.transformation} is then applied
to the normalized count matrix.

The "\eqn{+0.5}" and "\eqn{+1}" are taken from Law et al 2014,
and dropped from the normalization 
when the transformation is something else than \code{log2}.

The "\eqn{\times 10^6}" and "\eqn{\times 10^9}" factors are omitted when
the \code{asin(sqrt)} transformation is taken, as \eqn{asin} can only
be applied to real numbers smaller than 1.
}
\examples{
try(
if (require(ape) && require(evemodel)) {
tmpdir <- normalizePath(tempdir(), winslash = "/")
set.seed(20200317)
tree <- ape::read.tree(
  text = "(((A1:0,A2:0,A3:0):1,B1:1):1,((C1:0,C2:0):1.5,(D1:0,D2:0):1.5):0.5);")
mydata.obj <- generateSyntheticData(dataset = "mydata", n.vars = 100, 
                                    samples.per.cond = 4, n.diffexp = 10, 
                                    tree = tree,
                                    id.species = factor(c("A", "A", "A", "B", "C", "C", "D", "D")),
                                    lengths.relmeans = rpois(100, 1000),
                                    lengths.dispersions = rgamma(100, 1, 1),
                                    output.file = file.path(tmpdir, "mydata.rds"))
## Diff Exp
runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), result.extent = "twoThetaTest", 
           Rmdfunction = "evemodel.twoThetaTest.createRmd", 
           output.directory = tmpdir,
           norm.method = "TMM",
           length.normalization = "RPKM")
generateCodeHTMLs(file.path(tmpdir, "mydata_twoThetaTest.rds"), tmpdir)
\dontrun{
file.show(file.path(tmpdir, "mydata_twoThetaTest_Code.html"))
}
})
}
\references{
ROHLFS, Rori V. & NIELSEN, Rasmus. Phylogenetic ANOVA: The Expression Variance and Evolution Model for Quantitative Trait Evolution. 2015, 695–708. (64). ISSN: 1063-5157.

Law, C.W., Chen, Y., Shi, W. et al. (2014) voom: precision weights unlock linear model analysis tools for RNA-seq read counts. Genome Biol 15, R29.

Musser, JM, Wagner, GP. (2015): Character trees from transcriptome data: Origin and individuation of morphological characters and the so‐called “species signal”. J. Exp. Zool. (Mol. Dev. Evol.) 324B: 588– 604.
}
\author{
Charlotte Soneson, Paul Bastide, Mélina Gallopin
}
