% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhyloPoissonLogNormal.R
\name{get_model_factor}
\alias{get_model_factor}
\title{Get the scaling factor}
\usage{
get_model_factor(model.process, selection.strength, tree)
}
\arguments{
\item{model.process}{the process to be used for phylogenetic simulations. One of "BM" or "OU", default to "BM".}

\item{selection.strength}{if the process is "OU", the selection strength parameter for each gene. It can be a scalar, in which case the same parameter is used for all genes. Otherwise it needs to be a vector with length \code{n.vars}. Default to 0.}

\item{tree}{a dated phylogenetic tree of class \code{\link[ape]{phylo}}. If \code{samples.per.cond} is specified, it must have `samples.per.cond * 2` species.}
}
\value{
A vector N of factors.
}
\description{
Get the scaling factors.
}
\keyword{internal}
