% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{phyloCompData-class}
\alias{phyloCompData-class}
\title{Class phyloCompData}
\description{
The \code{phyloCompData} class extends the \code{\link{compData}} class
with sequence length and phylogeny related information.
}
\section{Slots}{

\describe{
\item{\code{tree}:}{The phylogenetic tree describing the relationships between samples. Class \code{phylo}}
\item{\code{length.matrix}:}{The length matrix, with genes as rows and samples as columns.  Class \code{matrix}}
\item{\code{sample.annotations}:}{Should contain an additional column named \code{id.species} of factors giving the species for each sample. If a tree is used, should have names matching the taxa of the tree.}
}
}

\section{Methods}{

\describe{
\item{phylo.tree}{\code{signature(x="compData")}}
\item{phylo.tree<-}{\code{signature(x="compData",value="phylo")}:
Get or set the tree in a \code{compData} object. \code{value} should be a phylo object.}
\item{length.matrix}{\code{signature(x="compData")}}
\item{length.matrix<-}{\code{signature(x="compData",value="matrix")}:
Get or set the length matrix in a \code{compData} object. \code{value} should be a numeric matrix.}
}
}

\section{Construction}{

An object of the class \code{phyloCompData} can be constructed using the \code{\link{phyloCompData}} function.
}

\author{
Charlotte Soneson, Paul Bastide
}
