% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylolmMethods.R
\name{phylolm_analysis}
\alias{phylolm_analysis}
\title{Perform the phylolm analysis}
\usage{
phylolm_analysis(
  dat,
  design_data,
  design_formula,
  tree,
  model,
  measurement_error,
  ...
)
}
\arguments{
\item{dat}{the data associated with a gene}

\item{design_data}{design matrix}

\item{design_formula}{design formula}

\item{tree}{phylogenetic tree}

\item{model}{the model to be used in phylolm}

\item{measurement_error}{boolean}
}
\value{
A list, with:
\describe{
\item{pvalue}{the p value of the differential expression.}
\item{logFC}{the log fold change of the differential expression.}
\item{score}{1 - pvalue.}
}
}
\description{
Perform the phylolm analysis for a given gene.
}
\keyword{internal}
