% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhyloPoissonLogNormal.R
\name{scale_variance_process}
\alias{scale_variance_process}
\title{Scale the variances}
\usage{
scale_variance_process(
  log_variance_phylo,
  tree,
  model.process,
  selection.strength
)
}
\arguments{
\item{tree}{a dated phylogenetic tree of class \code{\link[ape]{phylo}}. If \code{samples.per.cond} is specified, it must have `samples.per.cond * 2` species.}

\item{model.process}{the process to be used for phylogenetic simulations. One of "BM" or "OU", default to "BM".}

\item{selection.strength}{if the process is "OU", the selection strength parameter for each gene. It can be a scalar, in which case the same parameter is used for all genes. Otherwise it needs to be a vector with length \code{n.vars}. Default to 0.}
}
\value{
A matrix N * P of factors to multiply the simulated phylogenetic residuals.
}
\description{
Scale the variances of the process simulation so that they are equal to log_variance_phylo.
}
\keyword{internal}
