% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show,phyloCompData-method}
\alias{show,phyloCompData-method}
\title{Show method for \code{phyloCompData} object}
\usage{
\S4method{show}{phyloCompData}(object)
}
\arguments{
\item{object}{A \code{phyloCompData} object}
}
\description{
Show method for \code{phyloCompData} object.
}
\examples{
mydata <- generateSyntheticData(dataset = "mydata", n.vars = 1000, 
                                samples.per.cond = 5, n.diffexp = 100,
                                id.species = factor(1:10),
                                tree = ape::rphylo(10, 1, 0),
                                lengths.relmeans = "auto", lengths.dispersions = "auto")
mydata
}
\author{
Charlotte Soneson, Paul Bastide
}
