% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticData.R
\name{simulateData}
\alias{simulateData}
\title{Simulate the Data}
\usage{
simulateData(count_means, count_dispertions, overdispersed)
}
\arguments{
\item{count_means}{a matrix with the number of genes p rows and the number of
samples n columns.}

\item{count_dispertions}{a matrix of size p x n, for each gene and sample.}

\item{overdispersed}{Indices that are overdispersed.}
}
\value{
Z a n.var times 2*samples.per.cond matrix with the simulated data.
}
\description{
Use the Poisson or Negative Binomial model to simulate the data.
}
\keyword{internal}
