% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhyloPoissonLogNormal.R
\name{simulateDataPhylo}
\alias{simulateDataPhylo}
\title{Simulate the Data using the tree}
\usage{
simulateDataPhylo(count_means, count_dispertions, tree, prop.var.tree)
}
\arguments{
\item{count_means}{a matrix with the number of genes p rows and the number of
samples n columns.}

\item{count_dispertions}{a matrix of size p x n, for each gene and sample.}

\item{tree}{a phylogenetic tree of class \code{\link[ape:phylo]{phylo}} with `samples.per.cond * 2` species.}

\item{prop.var.tree}{the proportion of the common variance explained by the tree. Default to 1.}
}
\value{
Z a matrix with the data
}
\description{
Use the Phylogenetic Poisson Log Normal model to simulate the data.
}
\keyword{internal}
