% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhyloPoissonLogNormal.R
\name{simulateDataPhylo}
\alias{simulateDataPhylo}
\title{Simulate the Data using the tree}
\usage{
simulateDataPhylo(
  count_means,
  count_dispersions,
  tree,
  prop.var.tree,
  model.process = "BM",
  selection.strength = 0
)
}
\arguments{
\item{tree}{a dated phylogenetic tree of class \code{\link[ape]{phylo}}. If \code{samples.per.cond} is specified, it must have `samples.per.cond * 2` species.}

\item{prop.var.tree}{the proportion of the common variance explained by the tree for each gene. It can be a scalar, in which case the same parameter is used for all genes. Otherwise it needs to be a vector with length \code{n.vars}. Default to 1.}

\item{model.process}{the process to be used for phylogenetic simulations. One of "BM" or "OU", default to "BM".}

\item{selection.strength}{if the process is "OU", the selection strength parameter for each gene. It can be a scalar, in which case the same parameter is used for all genes. Otherwise it needs to be a vector with length \code{n.vars}. Default to 0.}
}
\value{
Z a matrix with the data
}
\description{
Use the Phylogenetic Poisson Log Normal model to simulate the data.
}
\keyword{internal}
